v3.5.0 Release
Changelog:
-Improved support for JPEGs and edit lists.
-Bug fixes, minor improvements and code cleanup.

+++ srcs/api-java/cpp/OutputStream.cpp
@@ -1,14 +1,13 @@
 /*
  * This file is part of Nokia HEIF library
  *
- * Copyright (c) 2015-2018 Nokia Corporation and/or its subsidiary(-ies). All rights reserved.
+ * Copyright (c) 2015-2019 Nokia Corporation and/or its subsidiary(-ies). All rights reserved.
  *
  * Contact: heif@nokia.com
  *
  * This software, including documentation, is protected by copyright controlled by Nokia Corporation and/ or its
  * subsidiaries. All rights are reserved. Copying, including reproducing, storing, adapting or translating, any or all
  * of this material requires the prior written consent of Nokia.
- *
  *
  */
 #include "OutputStream.h"
@@ -39,7 +38,7 @@ OutputStream::~OutputStream()
 
 void OutputStream::write(const void* buffer, std::uint64_t size)
 {
-    jobject byteBuffer = mJNIEnv->NewDirectByteBuffer(const_cast<void*>(buffer), size);
+    jobject byteBuffer = mJNIEnv->NewDirectByteBuffer(const_cast<void*>(buffer), static_cast<jlong>(size));
     mJNIEnv->CallVoidMethod(mJavaStream, mWriteMethodId, byteBuffer, size);
 }
